ASoC: mchp-spdifrx: add support for direct path
Add support for direct path on SPDIFRX. With this SPDIFRX IP could TX
data directly to ASRC when using in DPCM configs with out the need of a
DMA channel.

[ryan.wanner@microchip.com: Port changes into v6.6 kernel and adjust
naming to match the current naming of the SoC. Clean up commit message
to make it more clear. Add ability to disable direct path mode via DT
property.]

Signed-off-by: Claudiu Beznea <claudiu.beznea@microchip.com>
Signed-off-by: Ryan Wanner <Ryan.Wanner@microchip.com>