ASoC: mchp-spdiftx: add support for direct path
Add support for direct path on SPDIFTX. With this SPDIFTX IP could RX
data directly from ASRC when used in DPCM configs w/o the need of a
DMA channel.

[ryan.wanner@microchip.com: Port changes into v6.6 kernel. Adjust device
naming to match the current naming of the SoC. Add a check to disable
IRQs if direct path mode is enabled since they are not used for direct
path mode. Add ability to disable direct path mode via DT.]

Signed-off-by: Claudiu Beznea <claudiu.beznea@microchip.com>
Signed-off-by: Ryan Wanner <Ryan.Wanner@microchip.com>