ASoC: atmel: mchp-pdmc: add support for direct path
Add support for direct path on PDMC. With this PDMC IP could TX
data directly to ASRC when used in DPCM configs w/o the need of a
DMA channel.

[ryan.wanner@microchip.com: Port changes into v6.6 kernel. Disable IRQs
when using direct path mode since those are not needed in direct path
mode. Adjust naming of the SoC to match current naming of the SoC. Add
the ability to disable direct path mode via device tree.]

Signed-off-by: Claudiu Beznea <claudiu.beznea@microchip.com>
Signed-off-by: Ryan Wanner <Ryan.Wanner@microchip.com>