mmc: sdhci-of-at91: update clock management on pm runtime
Sama7 boards are impacted by an unbalanced prepare/unprepare
and enable/disable clock management in sdhci driver due to
the lack of support for SDMMC clock runtime disabling,
managed instead by the PMC driver.

Add conditions in the preset function to manage clock
depending on if the board support runtime clock disabling
and if we come from the probing or restoring procedure.

Add boolean start_clks variable to manage condition
in preset function.

Signed-off-by: Romain Sioen <romain.sioen@microchip.com>
Acked-by: Aubin Constans <aubin.constans@microchip.com>